created simple detail ui and moved screens into their own directories
