+++ app/shared/src/commonMain/kotlin/io/github/tscholze/cmpsample/RootContent.kt
@@ -1,7 +1,7 @@
 package io.github.tscholze.cmpsample
 
 import androidx.compose.runtime.Composable
-import io.github.tscholze.cmpsample.composables.screens.LocalResourceScreen
+import io.github.tscholze.cmpsample.composables.screens.list.LocalResourceScreen
 
 /**
  * Entry point of the app.
